api/python/requirements.txt
openai-agents
pytest
python-dotenv
pytest-asyncio>=0.20.0